Adding DeviceForm window
DeviceForm window is for adding new devices. The program first gets data from the Device.psa, if already exist and binding it to DataGridView. If the file is not presented the operator gets eror message. After adding new device by clicking on the button the application serialize the new entry and places it on the DataGridView.