Done with the Price acceptance, repair complete and device takeover
Price acceptance button is clicked after the customer accepts or not the repair cost. After that event the FixOrNot variable is changed to true or false. Repair complete is for completing the maintance. This method writes the date to the RepairFinishDate. The Device takover is when the customer gets the repaired device, fills the OutDate variable.